Add package exports and rename UpdatePasswordForm > ChangePasswordForm
