+++ src/index.js
@@ -1,5 +1,7 @@
 export { default as Copyright } from "./Copyright";
+export { default as ChangePasswordForm } from "./ChangePasswordForm";
 export { default as LoginForm } from "./LoginForm";
 export { default as RegistrationForm } from "./RegistrationForm";
 export { default as Select } from "./Select";
 export { default as ThemeProvider } from "./ThemeProvider";
+export { default as UserMenu } from "./UserMenu";
